Fix ingredient create/insert, modify dbhandler instantiation
Fixed the ingredient insertion statement to accept the correct number of values.
Corrected the ingredient table creation to store price as text rather than an integer.
Modified DatabaseHandler to accept a database name as a parameter to custom set the table name, so as not to interfere with live data when testing.