Add recipe database access, with tests
Add functionality to retrieve recipes from the database, and save to the database. CURRENTLY NOT INTEGRATED INTO THE MAIN SYSTEM. The way we access the database will switch to this once I have all the tables set up to interact.

Add hashcode and equals overrides to ingredient, and getOrAdd method to get or add an ingredient from the database.